chore: add faqs, case study, opentelemetry to blogs
